Move cleanup code into a separate module
There will be very much more as time goes on.

Use newly created wipe_blockdevs() method where appropriate.

Tidy up some cleanup code adding TODOs where appropriate.
